+++ src/ds/app/engine/engine_io.cpp
@@ -41,7 +41,7 @@ EngineSender::AutoSend::~AutoSend() {
 	if (!mSender.mConnection.initialized()) return;
 	if (mData.size() < 1) return;
 
-	const size_t size = mData.size();
+	const int size = static_cast<int>(mData.size());
 	mSender.mRawDataBuffer.setSize(size);
 	mData.readRaw(mSender.mRawDataBuffer.data(), size);
 	snappy::Compress(mSender.mRawDataBuffer.data(), size, &mSender.mCompressionBuffer);
@@ -135,7 +135,7 @@ bool EngineReceiver::handleBlob(ds::BlobRegistry& registry, ds::BlobReader& read
 	mNoDataCount = 0;
 
 	mCurrentDataBuffer.clear(); 
-	mCurrentDataBuffer.addRaw(mReceiveBuffers.front().c_str(), mReceiveBuffers.front().size());
+	mCurrentDataBuffer.addRaw(mReceiveBuffers.front().c_str(), static_cast<unsigned int>(mReceiveBuffers.front().size()));
 	mReceiveBuffers.erase(mReceiveBuffers.begin());
 
 	morePacketsAvailable = !mReceiveBuffers.empty();
